[Test framework MVP] RealmLocalizationResourceTest
